requirements.txt
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
requests
moviepy==1.0.3
Pillow
google-genai
google-cloud-secret-manager
